+++ toolium/test/pageelements/test_page_element.py
@@ -15,6 +15,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
+import os
 
 import mock
 import pytest
@@ -22,23 +23,34 @@ from selenium.common.exceptions import NoSuchElementException, TimeoutException
 from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
 
+from toolium.config_files import ConfigFiles
 from toolium.driver_wrapper import DriverWrapper
 from toolium.driver_wrappers_pool import DriverWrappersPool
-from toolium.pageelements import PageElement
+from toolium.pageelements import PageElement, Group
 from toolium.pageobjects.page_object import PageObject
 
 child_element = 'child_element'
 mock_element = None
 
 
+class MenuGroup(Group):
+    logo = PageElement(By.ID, 'image')
+    logo_wait = PageElement(By.ID, 'image2', wait=True)
+
+
 class RegisterPageObject(PageObject):
     username = PageElement(By.XPATH, '//input[0]')
     password = PageElement(By.ID, 'password', username)
+    menu_group = MenuGroup(By.ID, 'menu')
 
     def init_page_elements(self):
         self.language = PageElement(By.ID, 'language')
         self.email = PageElement(By.ID, 'email', mock_element)
         self.address = PageElement(By.ID, 'address', (By.ID, 'parent'))
+        self.address_shadowroot = PageElement(By.CSS_SELECTOR, '#address', shadowroot='shadowroot_css')
+        self.address_shadowroot_by_id = PageElement(By.ID, 'address', shadowroot='shadowroot_css')
+        self.element_webview = PageElement(By.ID, 'webview', webview=True)
+
 
 
 @pytest.fixture
@@ -54,7 +66,18 @@ def driver_wrapper():
 
     # Create a new wrapper
     driver_wrapper = DriverWrappersPool.get_default_wrapper()
+
+    # Configure wrapper
+    root_path = os.path.dirname(os.path.realpath(__file__))
+    config_files = ConfigFiles()
+    config_files.set_config_directory(os.path.join(root_path, 'conf'))
+    config_files.set_output_directory(os.path.join(root_path, 'output'))
+    config_files.set_config_log_filename('logging.conf')
+    DriverWrappersPool.configure_common_directories(config_files)
+    driver_wrapper.configure_properties()
+
     driver_wrapper.driver = mock.MagicMock()
+    driver_wrapper.is_mobile_test = mock.MagicMock(return_value=False)
 
     return driver_wrapper
 
@@ -137,6 +160,21 @@ def test_get_web_element_with_parent_web_element(driver_wrapper):
     mock_element.find_element.assert_called_once_with(By.ID, 'email')
 
 
+def test_get_web_element_shadowroot(driver_wrapper):
+    RegisterPageObject(driver_wrapper).address_shadowroot.web_element
+    expected_script = 'return document.querySelector("shadowroot_css").shadowRoot.querySelector("#address")'
+
+    mock_element.find_element.assert_not_called()
+    driver_wrapper.driver.execute_script.assert_called_once_with(expected_script)
+
+
+def test_get_web_element_shadowroot_wrong_locator(driver_wrapper):
+    with pytest.raises(Exception) as excinfo:
+        RegisterPageObject(driver_wrapper).address_shadowroot_by_id.web_element
+    assert "Locator type should be CSS_SELECTOR using shadowroot but found: id" in str(excinfo.value)
+    mock_element.find_element.assert_not_called()
+
+
 def test_get_web_element_in_test(driver_wrapper):
     PageElement(By.XPATH, '//input[0]').web_element
 
@@ -144,9 +182,7 @@ def test_get_web_element_in_test(driver_wrapper):
 
 
 def test_get_web_element_two_times_saving_enabled(driver_wrapper):
-    # Mock Driver.save_web_element = True
-    driver_wrapper.config = mock.MagicMock()
-    driver_wrapper.config.getboolean_optional.return_value = True
+    driver_wrapper.config.set('Driver', 'save_web_element', 'true')
     login_page = RegisterPageObject(driver_wrapper)
     login_page.username.web_element
     login_page.username.web_element
@@ -156,9 +192,7 @@ def test_get_web_element_two_times_saving_enabled(driver_wrapper):
 
 
 def test_get_web_element_two_times_saving_disabled(driver_wrapper):
-    # Mock Driver.save_web_element = False
-    driver_wrapper.config = mock.MagicMock()
-    driver_wrapper.config.getboolean_optional.return_value = False
+    driver_wrapper.config.set('Driver', 'save_web_element', 'false')
     login_page = RegisterPageObject(driver_wrapper)
     login_page.username.web_element
     login_page.username.web_element
@@ -352,3 +386,81 @@ def test_get_attribute(driver_wrapper):
     RegisterPageObject(driver_wrapper).username.get_attribute('attribute_name')
 
     mock_element.get_attribute.assert_called_once_with('attribute_name')
+
+
+def test_automatic_context_selection_group(driver_wrapper):
+    driver_wrapper.utils.wait_until_element_visible = mock.MagicMock(return_value=mock_element)
+    driver_wrapper.is_mobile_test = mock.MagicMock(return_value=True)
+    driver_wrapper.config.set('Driver', 'automatic_context_selection', 'true')
+    driver_wrapper.driver.context = "NATIVE_APP"
+    driver_wrapper.driver.contexts = ["NATIVE_APP", "WEBVIEW"]
+
+    RegisterPageObject(driver_wrapper).menu_group.web_element
+    driver_wrapper.driver.switch_to.context.assert_not_called()
+    driver_wrapper.driver.find_element.assert_called_once_with(By.ID, 'menu')
+
+
+def test_automatic_context_selection_no_webview_context_available(driver_wrapper):
+    driver_wrapper.is_mobile_test = mock.MagicMock(return_value=True)
+    driver_wrapper.config = mock.MagicMock()
+    driver_wrapper.config.set('Driver', 'automatic_context_selection', 'true')
+    driver_wrapper.driver.context = "NATIVE_APP"
+    driver_wrapper.driver.contexts = ["NATIVE_APP"]
+
+    with pytest.raises(Exception) as excinfo:
+        RegisterPageObject(driver_wrapper).element_webview.web_element
+    assert "WEBVIEW context not found" in str(excinfo.value)
+
+
+def test_automatic_context_selection_native_to_webview(driver_wrapper):
+    driver_wrapper.is_mobile_test = mock.MagicMock(return_value=True)
+    driver_wrapper.config.set('Driver', 'automatic_context_selection', 'true')
+    driver_wrapper.driver.context = "NATIVE_APP"
+    driver_wrapper.driver.contexts = ["NATIVE_APP", "WEBVIEW"]
+
+    RegisterPageObject(driver_wrapper).element_webview.web_element
+    driver_wrapper.driver.switch_to.context.assert_called_once_with("WEBVIEW")
+    driver_wrapper.driver.find_element.assert_called_once_with(By.ID, 'webview')
+
+
+def test_automatic_context_selection_webview_to_native(driver_wrapper):
+    driver_wrapper.is_mobile_test = mock.MagicMock(return_value=True)
+    driver_wrapper.config.set('Driver', 'automatic_context_selection', 'true')
+    driver_wrapper.driver.context = "WEBVIEW"
+    driver_wrapper.driver.contexts = ["NATIVE_APP", "WEBVIEW"]
+
+    RegisterPageObject(driver_wrapper).language.web_element
+    driver_wrapper.driver.switch_to.context.assert_called_once_with("NATIVE_APP")
+    driver_wrapper.driver.find_element.assert_called_once_with(By.ID, 'language')
+
+
+def test_automatic_context_selection_native_to_native(driver_wrapper):
+    driver_wrapper.is_mobile_test = mock.MagicMock(return_value=True)
+    driver_wrapper.config.set('Driver', 'automatic_context_selection', 'true')
+    driver_wrapper.driver.context = "NATIVE_APP"
+    driver_wrapper.driver.contexts = ["NATIVE_APP", "WEBVIEW"]
+
+    RegisterPageObject(driver_wrapper).language.web_element
+    driver_wrapper.driver.switch_to.context.assert_not_called()
+    driver_wrapper.driver.find_element.assert_called_once_with(By.ID, 'language')
+
+
+def test_automatic_context_selection_webview_to_webview(driver_wrapper):
+    driver_wrapper.is_mobile_test = mock.MagicMock(return_value=True)
+    driver_wrapper.config.set('Driver', 'automatic_context_selection', 'true')
+    driver_wrapper.driver.context = "WEBVIEW"
+    driver_wrapper.driver.contexts = ["NATIVE_APP", "WEBVIEW"]
+
+    RegisterPageObject(driver_wrapper).element_webview.web_element
+    driver_wrapper.driver.switch_to.context.assert_not_called()
+    driver_wrapper.driver.find_element.assert_called_once_with(By.ID, 'webview')
+
+
+def test_automatic_context_selection_disabled(driver_wrapper):
+    driver_wrapper.is_mobile_test = mock.MagicMock(return_value=True)
+    driver_wrapper.config.set('Driver', 'automatic_context_selection', 'false')
+    PageElement._automatic_context_selection = mock.MagicMock()
+
+    RegisterPageObject(driver_wrapper).element_webview.web_element
+    PageElement._automatic_context_selection.assert_not_called()
+    driver_wrapper.driver.find_element.assert_called_once_with(By.ID, 'webview')