+++ toolium/test/test_config_driver.py
@@ -35,9 +35,17 @@ def config():
     return config_parser
 
 
-def test_create_driver_local_not_configured(config):
+@pytest.fixture
+def utils():
+    utils = mock.MagicMock()
+    utils.get_driver_name.return_value = 'firefox'
+    return utils
+
+
+def test_create_driver_local_not_configured(config, utils):
     config.set('Driver', 'type', 'firefox')
-    config_driver = ConfigDriver(config)
+    utils.get_driver_name.return_value = 'firefox'
+    config_driver = ConfigDriver(config, utils)
     config_driver._create_local_driver = lambda: 'local driver mock'
     config_driver._create_remote_driver = lambda: 'remote driver mock'
 
@@ -46,10 +54,11 @@ def test_create_driver_local_not_configured(config):
     assert driver == 'local driver mock'
 
 
-def test_create_driver_local(config):
+def test_create_driver_local(config, utils):
     config.set('Server', 'enabled', 'false')
     config.set('Driver', 'type', 'firefox')
-    config_driver = ConfigDriver(config)
+    utils.get_driver_name.return_value = 'firefox'
+    config_driver = ConfigDriver(config, utils)
     config_driver._create_local_driver = lambda: 'local driver mock'
     config_driver._create_remote_driver = lambda: 'remote driver mock'
 
@@ -58,10 +67,11 @@ def test_create_driver_local(config):
     assert driver == 'local driver mock'
 
 
-def test_create_driver_remote(config):
+def test_create_driver_remote(config, utils):
     config.set('Server', 'enabled', 'true')
     config.set('Driver', 'type', 'firefox')
-    config_driver = ConfigDriver(config)
+    utils.get_driver_name.return_value = 'firefox'
+    config_driver = ConfigDriver(config, utils)
     config_driver._create_local_driver = lambda: 'local driver mock'
     config_driver._create_remote_driver = lambda: 'remote driver mock'
 
@@ -72,11 +82,12 @@ def test_create_driver_remote(config):
 
 @mock.patch('toolium.config_driver.Options')
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_local_driver_firefox(webdriver_mock, options, config):
+def test_create_local_driver_firefox(webdriver_mock, options, config, utils):
     config.set('Driver', 'type', 'firefox')
     config.add_section('Capabilities')
     config.set('Capabilities', 'marionette', 'false')
-    config_driver = ConfigDriver(config)
+    utils.get_driver_name.return_value = 'firefox'
+    config_driver = ConfigDriver(config, utils)
     config_driver._create_firefox_profile = lambda: 'firefox profile'
     DriverWrappersPool.output_directory = ''
 
@@ -90,12 +101,13 @@ def test_create_local_driver_firefox(webdriver_mock, options, config):
 
 @mock.patch('toolium.config_driver.Options')
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_local_driver_firefox_gecko(webdriver_mock, options, config):
+def test_create_local_driver_firefox_gecko(webdriver_mock, options, config, utils):
     config.set('Driver', 'type', 'firefox')
     config.add_section('Capabilities')
     config.set('Capabilities', 'marionette', 'true')
     config.set('Driver', 'gecko_driver_path', '/tmp/driver')
-    config_driver = ConfigDriver(config)
+    utils.get_driver_name.return_value = 'firefox'
+    config_driver = ConfigDriver(config, utils)
     config_driver._create_firefox_profile = lambda: 'firefox profile'
     DriverWrappersPool.output_directory = ''
 
@@ -108,13 +120,14 @@ def test_create_local_driver_firefox_gecko(webdriver_mock, options, config):
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_local_driver_firefox_binary(webdriver_mock, config):
+def test_create_local_driver_firefox_binary(webdriver_mock, config, utils):
     config.set('Driver', 'type', 'firefox')
     config.add_section('Capabilities')
     config.set('Capabilities', 'marionette', 'false')
     config.add_section('Firefox')
     config.set('Firefox', 'binary', '/tmp/firefox')
-    config_driver = ConfigDriver(config)
+    utils.get_driver_name.return_value = 'firefox'
+    config_driver = ConfigDriver(config, utils)
     config_driver._create_firefox_profile = lambda: 'firefox profile'
     DriverWrappersPool.output_directory = ''
 
@@ -131,10 +144,11 @@ def test_create_local_driver_firefox_binary(webdriver_mock, config):
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_local_driver_chrome(webdriver_mock, config):
+def test_create_local_driver_chrome(webdriver_mock, config, utils):
     config.set('Driver', 'type', 'chrome')
     config.set('Driver', 'chrome_driver_path', '/tmp/driver')
-    config_driver = ConfigDriver(config)
+    utils.get_driver_name.return_value = 'chrome'
+    config_driver = ConfigDriver(config, utils)
     config_driver._create_chrome_options = lambda: 'chrome options'
 
     config_driver._create_local_driver()
@@ -143,19 +157,21 @@ def test_create_local_driver_chrome(webdriver_mock, config):
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_local_driver_safari(webdriver_mock, config):
+def test_create_local_driver_safari(webdriver_mock, config, utils):
     config.set('Driver', 'type', 'safari')
-    config_driver = ConfigDriver(config)
+    utils.get_driver_name.return_value = 'safari'
+    config_driver = ConfigDriver(config, utils)
 
     config_driver._create_local_driver()
     webdriver_mock.Safari.assert_called_once_with(desired_capabilities=DesiredCapabilities.SAFARI)
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_local_driver_opera(webdriver_mock, config):
+def test_create_local_driver_opera(webdriver_mock, config, utils):
     config.set('Driver', 'type', 'opera')
     config.set('Driver', 'opera_driver_path', '/tmp/driver')
-    config_driver = ConfigDriver(config)
+    utils.get_driver_name.return_value = 'opera'
+    config_driver = ConfigDriver(config, utils)
 
     config_driver._create_local_driver()
     webdriver_mock.Opera.assert_called_once_with(desired_capabilities=DesiredCapabilities.OPERA,
@@ -163,66 +179,73 @@ def test_create_local_driver_opera(webdriver_mock, config):
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_local_driver_iexplore(webdriver_mock, config):
+def test_create_local_driver_iexplore(webdriver_mock, config, utils):
     config.set('Driver', 'type', 'iexplore')
     config.set('Driver', 'explorer_driver_path', '/tmp/driver')
-    config_driver = ConfigDriver(config)
+    utils.get_driver_name.return_value = 'iexplore'
+    config_driver = ConfigDriver(config, utils)
 
     config_driver._create_local_driver()
     webdriver_mock.Ie.assert_called_once_with('/tmp/driver', capabilities=DesiredCapabilities.INTERNETEXPLORER)
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_local_driver_edge(webdriver_mock, config):
+def test_create_local_driver_edge(webdriver_mock, config, utils):
     config.set('Driver', 'type', 'edge')
     config.set('Driver', 'edge_driver_path', '/tmp/driver')
-    config_driver = ConfigDriver(config)
+    utils.get_driver_name.return_value = 'edge'
+    config_driver = ConfigDriver(config, utils)
 
     config_driver._create_local_driver()
     webdriver_mock.Edge.assert_called_once_with('/tmp/driver', capabilities=DesiredCapabilities.EDGE)
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_local_driver_phantomjs(webdriver_mock, config):
+def test_create_local_driver_phantomjs(webdriver_mock, config, utils):
     config.set('Driver', 'type', 'phantomjs')
     config.set('Driver', 'phantomjs_driver_path', '/tmp/driver')
-    config_driver = ConfigDriver(config)
+    utils.get_driver_name.return_value = 'phantomjs'
+    config_driver = ConfigDriver(config, utils)
 
     config_driver._create_local_driver()
     webdriver_mock.PhantomJS.assert_called_once_with(desired_capabilities=DesiredCapabilities.PHANTOMJS,
                                                      executable_path='/tmp/driver')
 
 
-def test_create_local_driver_android(config):
+def test_create_local_driver_android(config, utils):
     config.set('Driver', 'type', 'android')
-    config_driver = ConfigDriver(config)
+    utils.get_driver_name.return_value = 'android'
+    config_driver = ConfigDriver(config, utils)
     config_driver._create_remote_driver = lambda: 'remote driver mock'
 
     driver = config_driver._create_local_driver()
     assert driver == 'remote driver mock'
 
 
-def test_create_local_driver_ios(config):
+def test_create_local_driver_ios(config, utils):
     config.set('Driver', 'type', 'ios')
-    config_driver = ConfigDriver(config)
+    utils.get_driver_name.return_value = 'ios'
+    config_driver = ConfigDriver(config, utils)
     config_driver._create_remote_driver = lambda: 'remote driver mock'
 
     driver = config_driver._create_local_driver()
     assert driver == 'remote driver mock'
 
 
-def test_create_local_driver_iphone(config):
+def test_create_local_driver_iphone(config, utils):
     config.set('Driver', 'type', 'iphone')
-    config_driver = ConfigDriver(config)
+    utils.get_driver_name.return_value = 'iphone'
+    config_driver = ConfigDriver(config, utils)
     config_driver._create_remote_driver = lambda: 'remote driver mock'
 
     driver = config_driver._create_local_driver()
     assert driver == 'remote driver mock'
 
 
-def test_create_local_driver_unknown_driver(config):
+def test_create_local_driver_unknown_driver(config, utils):
     config.set('Driver', 'type', 'unknown')
-    config_driver = ConfigDriver(config)
+    utils.get_driver_name.return_value = 'unknown'
+    config_driver = ConfigDriver(config, utils)
 
     with pytest.raises(Exception) as excinfo:
         config_driver._create_local_driver()
@@ -231,12 +254,13 @@ def test_create_local_driver_unknown_driver(config):
 
 @mock.patch('toolium.config_driver.Options')
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_local_driver_capabilities(webdriver_mock, options, config):
+def test_create_local_driver_capabilities(webdriver_mock, options, config, utils):
     config.set('Driver', 'type', 'firefox')
     config.add_section('Capabilities')
     config.set('Capabilities', 'marionette', 'false')
     config.set('Capabilities', 'version', '45')
-    config_driver = ConfigDriver(config)
+    utils.get_driver_name.return_value = 'firefox'
+    config_driver = ConfigDriver(config, utils)
     config_driver._create_firefox_profile = lambda: 'firefox profile'
     DriverWrappersPool.output_directory = ''
 
@@ -250,11 +274,11 @@ def test_create_local_driver_capabilities(webdriver_mock, options, config):
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_remote_driver_firefox(webdriver_mock, config):
+def test_create_remote_driver_firefox(webdriver_mock, config, utils):
     config.set('Driver', 'type', 'firefox')
     server_url = 'http://10.20.30.40:5555'
-    utils = mock.MagicMock()
     utils.get_server_url.return_value = server_url
+    utils.get_driver_name.return_value = 'firefox'
     config_driver = ConfigDriver(config, utils)
 
     # Firefox profile mock
@@ -271,11 +295,11 @@ def test_create_remote_driver_firefox(webdriver_mock, config):
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_remote_driver_chrome(webdriver_mock, config):
+def test_create_remote_driver_chrome(webdriver_mock, config, utils):
     config.set('Driver', 'type', 'chrome')
     server_url = 'http://10.20.30.40:5555'
-    utils = mock.MagicMock()
     utils.get_server_url.return_value = server_url
+    utils.get_driver_name.return_value = 'chrome'
     config_driver = ConfigDriver(config, utils)
 
     # Chrome options mock
@@ -291,11 +315,11 @@ def test_create_remote_driver_chrome(webdriver_mock, config):
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_remote_driver_chrome_old_selenium(webdriver_mock, config):
+def test_create_remote_driver_chrome_old_selenium(webdriver_mock, config, utils):
     config.set('Driver', 'type', 'chrome')
     server_url = 'http://10.20.30.40:5555'
-    utils = mock.MagicMock()
     utils.get_server_url.return_value = server_url
+    utils.get_driver_name.return_value = 'chrome'
     config_driver = ConfigDriver(config, utils)
 
     # Chrome options mock
@@ -311,11 +335,11 @@ def test_create_remote_driver_chrome_old_selenium(webdriver_mock, config):
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_remote_driver_safari(webdriver_mock, config):
+def test_create_remote_driver_safari(webdriver_mock, config, utils):
     config.set('Driver', 'type', 'safari')
     server_url = 'http://10.20.30.40:5555'
-    utils = mock.MagicMock()
     utils.get_server_url.return_value = server_url
+    utils.get_driver_name.return_value = 'safari'
     config_driver = ConfigDriver(config, utils)
 
     config_driver._create_remote_driver()
@@ -324,11 +348,11 @@ def test_create_remote_driver_safari(webdriver_mock, config):
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_remote_driver_opera(webdriver_mock, config):
+def test_create_remote_driver_opera(webdriver_mock, config, utils):
     config.set('Driver', 'type', 'opera')
     server_url = 'http://10.20.30.40:5555'
-    utils = mock.MagicMock()
     utils.get_server_url.return_value = server_url
+    utils.get_driver_name.return_value = 'opera'
     config_driver = ConfigDriver(config, utils)
 
     config_driver._create_remote_driver()
@@ -340,11 +364,11 @@ def test_create_remote_driver_opera(webdriver_mock, config):
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_remote_driver_iexplore(webdriver_mock, config):
+def test_create_remote_driver_iexplore(webdriver_mock, config, utils):
     config.set('Driver', 'type', 'iexplore')
     server_url = 'http://10.20.30.40:5555'
-    utils = mock.MagicMock()
     utils.get_server_url.return_value = server_url
+    utils.get_driver_name.return_value = 'iexplore'
     config_driver = ConfigDriver(config, utils)
 
     config_driver._create_remote_driver()
@@ -353,11 +377,11 @@ def test_create_remote_driver_iexplore(webdriver_mock, config):
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_remote_driver_edge(webdriver_mock, config):
+def test_create_remote_driver_edge(webdriver_mock, config, utils):
     config.set('Driver', 'type', 'edge')
     server_url = 'http://10.20.30.40:5555'
-    utils = mock.MagicMock()
     utils.get_server_url.return_value = server_url
+    utils.get_driver_name.return_value = 'edge'
     config_driver = ConfigDriver(config, utils)
 
     config_driver._create_remote_driver()
@@ -366,11 +390,11 @@ def test_create_remote_driver_edge(webdriver_mock, config):
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_remote_driver_phantomjs(webdriver_mock, config):
+def test_create_remote_driver_phantomjs(webdriver_mock, config, utils):
     config.set('Driver', 'type', 'phantomjs')
     server_url = 'http://10.20.30.40:5555'
-    utils = mock.MagicMock()
     utils.get_server_url.return_value = server_url
+    utils.get_driver_name.return_value = 'phantomjs'
     config_driver = ConfigDriver(config, utils)
 
     config_driver._create_remote_driver()
@@ -379,14 +403,14 @@ def test_create_remote_driver_phantomjs(webdriver_mock, config):
 
 
 @mock.patch('toolium.config_driver.appiumdriver')
-def test_create_remote_driver_android(appiumdriver_mock, config):
+def test_create_remote_driver_android(appiumdriver_mock, config, utils):
     config.set('Driver', 'type', 'android')
     config.add_section('AppiumCapabilities')
     config.set('AppiumCapabilities', 'automationName', 'Appium')
     config.set('AppiumCapabilities', 'platformName', 'Android')
     server_url = 'http://10.20.30.40:5555'
-    utils = mock.MagicMock()
     utils.get_server_url.return_value = server_url
+    utils.get_driver_name.return_value = 'android'
     config_driver = ConfigDriver(config, utils)
 
     config_driver._create_remote_driver()
@@ -396,14 +420,14 @@ def test_create_remote_driver_android(appiumdriver_mock, config):
 
 
 @mock.patch('toolium.config_driver.appiumdriver')
-def test_create_remote_driver_ios(appiumdriver_mock, config):
+def test_create_remote_driver_ios(appiumdriver_mock, config, utils):
     config.set('Driver', 'type', 'ios')
     config.add_section('AppiumCapabilities')
     config.set('AppiumCapabilities', 'automationName', 'Appium')
     config.set('AppiumCapabilities', 'platformName', 'iOS')
     server_url = 'http://10.20.30.40:5555'
-    utils = mock.MagicMock()
     utils.get_server_url.return_value = server_url
+    utils.get_driver_name.return_value = 'ios'
     config_driver = ConfigDriver(config, utils)
 
     config_driver._create_remote_driver()
@@ -421,6 +445,7 @@ def test_create_remote_driver_iphone(appiumdriver_mock, config):
     server_url = 'http://10.20.30.40:5555'
     utils = mock.MagicMock()
     utils.get_server_url.return_value = server_url
+    utils.get_driver_name.return_value = 'iphone'
     config_driver = ConfigDriver(config, utils)
 
     config_driver._create_remote_driver()
@@ -430,11 +455,11 @@ def test_create_remote_driver_iphone(appiumdriver_mock, config):
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_remote_driver_version_platform(webdriver_mock, config):
+def test_create_remote_driver_version_platform(webdriver_mock, config, utils):
     config.set('Driver', 'type', 'iexplore-11-on-WIN10')
     server_url = 'http://10.20.30.40:5555'
-    utils = mock.MagicMock()
     utils.get_server_url.return_value = server_url
+    utils.get_driver_name.return_value = 'iexplore'
     config_driver = ConfigDriver(config, utils)
 
     config_driver._create_remote_driver()
@@ -446,11 +471,11 @@ def test_create_remote_driver_version_platform(webdriver_mock, config):
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_remote_driver_version(webdriver_mock, config):
+def test_create_remote_driver_version(webdriver_mock, config, utils):
     config.set('Driver', 'type', 'iexplore-11')
     server_url = 'http://10.20.30.40:5555'
-    utils = mock.MagicMock()
     utils.get_server_url.return_value = server_url
+    utils.get_driver_name.return_value = 'iexplore'
     config_driver = ConfigDriver(config, utils)
 
     config_driver._create_remote_driver()
@@ -461,13 +486,13 @@ def test_create_remote_driver_version(webdriver_mock, config):
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_remote_driver_capabilities(webdriver_mock, config):
+def test_create_remote_driver_capabilities(webdriver_mock, config, utils):
     config.set('Driver', 'type', 'iexplore-11')
     config.add_section('Capabilities')
     config.set('Capabilities', 'version', '11')
     server_url = 'http://10.20.30.40:5555'
-    utils = mock.MagicMock()
     utils.get_server_url.return_value = server_url
+    utils.get_driver_name.return_value = 'iexplore'
     config_driver = ConfigDriver(config, utils)
 
     config_driver._create_remote_driver()
@@ -477,45 +502,51 @@ def test_create_remote_driver_capabilities(webdriver_mock, config):
                                                   desired_capabilities=capabilities)
 
 
-def test_convert_property_type_true(config):
-    config_driver = ConfigDriver(config)
+def test_convert_property_type_true(config, utils):
+    config_driver = ConfigDriver(config, utils)
     value = 'True'
     assert config_driver._convert_property_type(value) is True
 
 
-def test_convert_property_type_false(config):
-    config_driver = ConfigDriver(config)
+def test_convert_property_type_false(config, utils):
+    config_driver = ConfigDriver(config, utils)
     value = 'False'
     assert config_driver._convert_property_type(value) is False
 
 
-def test_convert_property_type_dict(config):
-    config_driver = ConfigDriver(config)
+def test_convert_property_type_dict(config, utils):
+    config_driver = ConfigDriver(config, utils)
     value = "{'a': 5}"
     assert config_driver._convert_property_type(value) == {'a': 5}
 
 
-def test_convert_property_type_int(config):
-    config_driver = ConfigDriver(config)
+def test_convert_property_type_int(config, utils):
+    config_driver = ConfigDriver(config, utils)
     value = '5'
     assert config_driver._convert_property_type(value) == 5
 
 
-def test_convert_property_type_str(config):
-    config_driver = ConfigDriver(config)
+def test_convert_property_type_str(config, utils):
+    config_driver = ConfigDriver(config, utils)
     value = 'string'
     assert config_driver._convert_property_type(value) == value
 
 
+def test_convert_property_type_list(config, utils):
+    config_driver = ConfigDriver(config, utils)
+    value = "[1, 2, 3]"
+    assert config_driver._convert_property_type(value) == [1, 2, 3]
+
+
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_firefox_profile(webdriver_mock, config):
+def test_create_firefox_profile(webdriver_mock, config, utils):
     config.add_section('Firefox')
     config.set('Firefox', 'profile', '/tmp')
     config.add_section('FirefoxPreferences')
     config.set('FirefoxPreferences', 'browser.download.folderList', '2')
     config.add_section('FirefoxExtensions')
     config.set('FirefoxExtensions', 'firebug', 'resources/firebug-3.0.0-beta.3.xpi')
-    config_driver = ConfigDriver(config)
+    config_driver = ConfigDriver(config, utils)
 
     config_driver._create_firefox_profile()
     webdriver_mock.FirefoxProfile.assert_called_once_with(profile_directory='/tmp')
@@ -524,10 +555,10 @@ def test_create_firefox_profile(webdriver_mock, config):
     webdriver_mock.FirefoxProfile().add_extension.assert_called_once_with('resources/firebug-3.0.0-beta.3.xpi')
 
 
-def test_add_firefox_arguments(config):
+def test_add_firefox_arguments(config, utils):
     config.add_section('FirefoxArguments')
     config.set('FirefoxArguments', '-private', '')
-    config_driver = ConfigDriver(config)
+    config_driver = ConfigDriver(config, utils)
     firefox_options = Options()
 
     config_driver._add_firefox_arguments(firefox_options)
@@ -535,14 +566,14 @@ def test_add_firefox_arguments(config):
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_chrome_options(webdriver_mock, config):
+def test_create_chrome_options(webdriver_mock, config, utils):
     config.add_section('ChromePreferences')
     config.set('ChromePreferences', 'download.default_directory', '/tmp')
     config.add_section('ChromeMobileEmulation')
     config.set('ChromeMobileEmulation', 'deviceName', 'Google Nexus 5')
     config.add_section('ChromeArguments')
     config.set('ChromeArguments', 'lang', 'es')
-    config_driver = ConfigDriver(config)
+    config_driver = ConfigDriver(config, utils)
 
     config_driver._create_chrome_options()
     webdriver_mock.ChromeOptions.assert_called_once_with()
@@ -554,9 +585,9 @@ def test_create_chrome_options(webdriver_mock, config):
 
 
 @mock.patch('toolium.config_driver.webdriver')
-def test_create_chrome_options_headless(webdriver_mock, config):
+def test_create_chrome_options_headless(webdriver_mock, config, utils):
     config.set('Driver', 'headless', 'true')
-    config_driver = ConfigDriver(config)
+    config_driver = ConfigDriver(config, utils)
 
     config_driver._create_chrome_options()
     webdriver_mock.ChromeOptions.assert_called_once_with()